finalize migration
